Can get tags from Behance images
The problem with Behance is that it uses SSL, but Clarifai is having
issues with SSL, so it cannot use Behance images. However, it can still
get images from Imgur, so I made a service that will rehost images to
Imgur, so this file now will rehost to Imgur and then use the Imgur link
with Clarifai and that will allow for it to be usable with Behance.